Refactor (1) - Split components
* refactoring

* Refactor part 2

* Send props

* Refactor Delete Feature Toggle

* Rename

* Intermediary commit - edit feature toggle

* Updates

* Finish EditFeatureToggle.vue

* Finish EditFeatureToggle

* styling and update ft tests

* Refactor EditEnvironment & DeleteEnvironment

* Try fix Grid bug

* loadGridData before ResetGrid

* Update Toggle Schedule

* Rearrange files & start replacing events

* Fix grid error

* add events list
