+++ Moggles/ClientApp/src/boot.js
