+++ Moggles/ClientApp/src/common/events.js
@@ -1,0 +1,32 @@
+export const events = {
+    newApplicationAdded: "app-added",
+    applicationEdited: "app-edited",
+    reloadApplicationToggles: "reload-application-toggles",
+    refreshApplications: "refresh-apps",
+    openAddApplicationModal: "openAddAppModal",
+    closeAddApplicationModal: "close-add-application",
+    showDeleteApplicationConfirmationModal:"show-app-delete-confirmation",
+    applicationChanged: "app-changed",
+    environmentsLoaded: "env-loaded",
+    environmentAdded: "env-added",
+    openAddEnvironmentModal:"openAddEnvModal",
+    editEnvironment: "edit-environment",
+    closeAddEnvironmentModal: "close-add-environment",
+    closeDeleteEnvironmentModal: "close-deleteEnvironment",
+    closeEditEnvironmentModal: "close-editEnvironment",
+    togglesLoaded: "toggles-loaded",
+    openAddFeatureToggleModal: "openAddFeatureToggleModal",
+    toggleAdded: "toggle-added",
+    closeAddFeatureToggleModal: "close-add-toggle",
+    deleteFeatureToggle: "delete-featureToggle",
+    closeDeleteFeatureToggleModal: "close-deleteToggle",
+    openEditFeatureToggleModal: "open-editFeatureToggle",
+    closeEditFeatureToggleModal: "close-editFeatureToggle",
+    openAddSchedulerModal: "add-scheduler",
+    toggleScheduled: "toggle-scheduled",
+    closeToggleSchedulerModal: "close-scheduler",
+    closeDeleteSchedulerModal: "close-deleteScheduler",
+    closeForceCacheRefreshModal: "close-refresh",
+    unblockUI: "unblock-ui",
+    blockUI: "block-ui"
+}
