+++ Moggles/ClientApp/src/menu/AppSelection.vue
@@ -1,61 +1,69 @@
 ﻿<template>
   <div id="app-sel"> 
-    <multi-select ref="appSelection" v-model="selectedApps" :limit="1" 
-                  :options="applicationList" :value-key="'id'" :label-key="'appName'"
+    <multi-select ref="appSelection" v-model="selectedApp" :limit="1" 
+                  :options="applications" :value-key="'id'" :label-key="'appName'"
                   :selected-icon="'fas fa-check'" class="padding-left-10" @change="changeApp" />
   </div>
 </template>
 
 <script>
-	import { Bus } from './event-bus'
-	import axios from 'axios'
-	import _ from 'lodash'
+	import { Bus } from '../common/event-bus';
+	import { events } from '../common/events';
+	import axios from 'axios';
+	import _ from 'lodash';
 
 	export default {
 		data() {
 			return {
-				applicationList: [],
-				selectedApps: []
+				applications: [],
+				selectedApp: []
 			}
 		},
 		computed: {
 			appIsSelected() {
-				return this.selectedApps.length > 0;
+				return this.selectedApp.length > 0;
 			}
 		},
 		created() {
-			this.getApplications()
-			Bus.$on("new-app-added", () => {
+			this.getApplications();
+
+			Bus.$on(events.newApplicationAdded, () => {
 				this.getApplications();
 			});
-			Bus.$on("reload-application-toggles", () => {
+
+			Bus.$on(events.applicationEdited, () => {
+				this.getApplications();
+			});
+
+			Bus.$on(events.reloadApplicationToggles, () => {
 				this.changeApp();
 			});
-			Bus.$on("refresh-apps", () => {
+
+			Bus.$on(events.refreshApplications, () => {
 				this.refreshApps();
 			});
 		},
 		methods: {
             changeApp() {
-                var app =_.find(this.applicationList, (a) => a.id == this.selectedApps[0]);
+                var app =_.find(this.applications, (a) => a.id == this.selectedApp[0]);
 				if (app) {
-                    Bus.$emit('app-changed', app)
+                    Bus.$emit(events.applicationChanged, app);
                     localStorage.setItem('selectedApp', app.id);
-					this.$refs.appSelection.showDropdown = false
+					this.$refs.appSelection.showDropdown = false;
 				}
 			},
             getApplications() {
 				axios.get('/api/applications')
 					.then((response) => {
-						this.applicationList = response.data
+						this.applications = response.data
 						if (!this.appIsSelected) {
 							if (response.data.length > 0) {
-                                if (this.selectedApps.length == 0) {
+                                if (this.selectedApp.length == 0) {
                                     if (localStorage.getItem('selectedApp') === null || !this.existsStoredApp()) {
-                                        this.selectedApps.push(response.data[0].id);
+                                        this.selectedApp.push(response.data[0].id);
                                     }
                                     else {
-                                        this.selectedApps.push(localStorage.getItem('selectedApp'));
+                                        this.selectedApp.push(localStorage.getItem('selectedApp'));
                                     }
                                     this.changeApp();
 								}
@@ -65,18 +73,18 @@
 					.catch(error => { window.alert(error) })
 			},
 			refreshApps() {
-				this.selectedApps = []
-				this.getApplications()
+				this.selectedApp = [];
+				this.getApplications();
             },
             getAllAplications() {
                 axios.get('/api/applications')
                     .then((response) => {
-                        this.applicationList = response.data;
+                        this.application = response.data;
                     })
                     .catch(error => { window.alert(error) });
             },
             existsStoredApp() {
-                var app = _.find(this.applicationList, (a) => a.id == localStorage.getItem('selectedApp'));
+                var app = _.find(this.application, (a) => a.id == localStorage.getItem('selectedApp'));
                 return app != null 
             }
 		}
