+++ Moggles/ClientApp/src/menu/ForceCacheRefresh.vue
@@ -31,8 +31,9 @@
 </template>
 
 <script>
-	import { Bus } from './event-bus';
+	import { Bus } from '../common/event-bus';
 	import axios from 'axios';
+	import { events } from '../common/events';
 
 	export default {
 		data() {
@@ -46,13 +47,13 @@
 			};
 		},
 		mounted() {
-			Bus.$on("app-changed", app => {
+			Bus.$on(events.applicationChanged, app => {
 				if (app) {
 					this.applicationId = app.id;
 				}
 			});
 
-			Bus.$on("env-loaded", envs => {
+			Bus.$on(events.environmentsLoaded, envs => {
 				this.existingEnvs = envs;
 			});
 		},
@@ -66,19 +67,19 @@
 					envName: this.envName
 				};
 
-				Bus.$emit('block-ui')
+                Bus.$emit(events.blockUI)
 				axios.post('api/CacheRefresh', param)
 					.then(() => {
 						this.showSuccessAlert = true;
 						this.envName = null;
 					}).catch((e) => {
 						window.alert(e);
-					}).finally(() => {
-						Bus.$emit('unblock-ui')
+                    }).finally(() => {
+                        Bus.$emit(events.unblockUI)
 					});
 			},
 			closeRefreshModal() {
-				Bus.$emit('close-refresh');
+				Bus.$emit(events.closeForceCacheRefreshModal);
 			}
 		}
 	}
