+++ Moggles/Models/FeatureToggleUpdateModel.cs
@@ -15,7 +15,7 @@ namespace Moggles.Models
         public bool IsPermanent { get; set; }
         public List<FeatureToggleStatusUpdateModel> Statuses { get; set; }
         public string WorkItemIdentifier { get; set; }
-        public List<FeatureToggleReasonToChangeModel> ReasonsToChange { get; set; }
+        public string ReasonToChange { get; set; }
     }
 
     public class FeatureToggleStatusUpdateModel
@@ -23,10 +23,5 @@ namespace Moggles.Models
         public bool Enabled { get; set; }
         public string Environment { get; set; }
         public string UpdatedByUser { get; set; }
-    }
-
-    public class FeatureToggleReasonToChangeModel
-    {
-        public string Description { get; set; }
     }
 }
