--- Moggles/appsettings.json
@@ -1,15 +1,0 @@
-{
-  "Logging": {
-    "IncludeScopes": false,
-    "LogLevel": {
-      "Default": "Warning"
-    }
-  },
-  "CustomRoles": {
-  },
-  "Messaging": {
-  },
-  "ApplicationInsights": {
-    "InstrumentationKey": ""
-  }
-}
